+++ connect_infobip/__manifest__.py
@@ -1,6 +1,6 @@
 {
     'name': 'Oduist Connect Infobip',
-    'version': '19.0.1.2.0',
+    'version': '19.0.1.2.1',
     'author': 'Oduist',
     'category': 'Phone',
     'summary': 'Infobip integration for Oduist Connect',